+++ packages/types/CHANGELOG.md
@@ -8,7 +8,7 @@ Versioning].
 
 ## [Unreleased]
 
-- Add case sensitivity flag to `CssDeclarationValueOptions`. ([#338])
+- Add `kind` field to `CssDeclarationValueOptions`. ([#378])
 - Add `kind` field to `QuerySelectorOptions`. ([#366])
 
 ## [0.1.26] - 2022-07-02
@@ -38,7 +38,7 @@ Versioning].
 [#306]: https://github.com/ericcornelissen/webmangler/pull/306
 [#311]: https://github.com/ericcornelissen/webmangler/pull/311
 [#314]: https://github.com/ericcornelissen/webmangler/pull/314
-[#338]: https://github.com/ericcornelissen/webmangler/pull/338
 [#366]: https://github.com/ericcornelissen/webmangler/pull/366
+[#378]: https://github.com/ericcornelissen/webmangler/pull/378
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/ "Keep a CHANGELOG"
 [semantic versioning]: https://semver.org/spec/v2.0.0.html "Semantic versioning"